Add second part to project, update README
The Odin Project source lesson had moved on to a second part, so this
project has added a second part to keep pace. We needed a way to keep
both parts separate, so folders have been created for *part1* and
*part2*.

Also added the first activity to the part2 folder. No solution has been
attempted in the activity at this time.